requirements.txt
browser-use==0.1.19
langchain-openai==0.3.0
python-dotenv==1.0.1